django-railway-backend/requirements.txt
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
django-filter==24.3
psycopg2-binary==2.9.9
dj-database-url==2.1.0
Pillow==10.4.0
reportlab==4.2.5
python-dotenv==1.0.0
gunicorn==22.0.0
whitenoise==6.7.0
django-storages==1.14.4
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.147.0
